backend/requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.7.4,<3.0.0
pydantic-settings>=2.1.0

# LangGraph & LangChain
langgraph>=0.2.45
langchain>=0.3.7
langchain-core>=0.3.15
langchain-groq>=0.2.0

# Data Processing
pandas>=2.2.3
numpy>=1.26.4

# CORS & HTTP
python-multipart==0.0.6
python-dotenv==1.0.0

# Date/Time
python-dateutil==2.8.2
# Computer Vision & Object Detection
ultralytics>=8.1.0  # YOLOv8
opencv-python-headless>=4.8.0  # OpenCV for image processing (headless for server)
pillow>=10.0.0  # Image processing

# Real-time WebSocket
python-socketio>=5.10.0
python-engineio>=4.8.0